Added failing single lib same point
Added test case in which there are two tests for one function, of which
one passes and one fails, and the assumption is that these two tests are
required to recieve a point. So if one test fails and one passes you will
still not get a point because they both have to be right. There is also
one unrelated passing test for another function.